+++ app/src/main/java/app/urv/manager/patcher/runtime/process/Revanced22PatcherProcess.kt
@@ -193,6 +193,7 @@ class Revanced22PatcherProcess(
             put("outputFile", parameters.outputFile)
             put("stripNativeLibs", parameters.stripNativeLibs)
             put("skipUnneededSplits", parameters.skipUnneededSplits)
+            put("continueOnPatchError", parameters.continueOnPatchError)
             put("patcherLogMode", parameters.patcherLogMode.name)
             put("configurations", configs)
         }