+++ app/src/main/java/app/urv/manager/patcher/runtime/revanced/Revanced21BridgeRuntime.kt
@@ -69,6 +69,7 @@ class Revanced21BridgeRuntime(context: Context) : Runtime(context) {
             "outputFile" to outputFile,
             "stripNativeLibs" to stripNativeLibs,
             "skipUnneededSplits" to skipUnneededSplits,
+            "continueOnPatchError" to prefs.continueOnPatchError.get(),
             "patcherLogMode" to logMode.name,
             "configurations" to configs
         )