+++ app/src/main/java/app/urv/manager/patcher/runtime/revanced/Revanced21ProcessRuntime.kt
@@ -288,6 +288,7 @@ class Revanced21ProcessRuntime(
                 },
                 stripNativeLibs = stripNativeLibs,
                 skipUnneededSplits = skipUnneededSplits,
+                continueOnPatchError = prefs.continueOnPatchError.get(),
                 patcherLogMode = logMode
             )