+++ app/src/main/java/app/urv/manager/patcher/runtime/revanced/Revanced22BridgeRuntime.kt
@@ -132,6 +132,7 @@ class Revanced22BridgeRuntime(context: Context) : Runtime(context) {
                 "patcherLogMode" to logMode.name,
                 "stripNativeLibs" to stripNativeLibs,
                 "skipUnneededSplits" to skipUnneededSplits,
+                "continueOnPatchError" to prefs.continueOnPatchError.get(),
                 "configurations" to configs
             )